LayoutTests/fast/multicol/columns-shorthand-parsing-2-expected.txt

Test the behavior when 'auto' is part of the 'columns' property value. See http://www.w3.org/TR/css3-multicol/#columns

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "3"
PASS element.style.columns is "3"
PASS element.style.columnWidth is "10em"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "10em"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "3"
PASS element.style.columns is "3"
PASS element.style.columnWidth is "10em"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "10em"
PASS element.style.columnWidth is "7em"
PASS element.style.columnCount is "7"
PASS element.style.columns is "7em 7"
PASS element.style.columnWidth is "7em"
PASS element.style.columnCount is "7"
PASS element.style.columns is "7em 7"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "auto"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "auto"
PASS element.style.columnWidth is "initial"
PASS element.style.columnCount is "initial"
PASS element.style.columns is "initial"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "auto"
PASS element.style.columnWidth is "inherit"
PASS element.style.columnCount is "inherit"
PASS element.style.columns is "inherit"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "auto"
PASS element.style.columnWidth is "auto"
PASS element.style.columnCount is "7"
PASS element.style.columns is "7"
PASS element.style.columnWidth is "7em"
PASS element.style.columnCount is "auto"
PASS element.style.columns is "7em"
PASS element.style.columnWidth is "0px"
PASS element.style.columnCount is "1"
PASS element.style.columns is "0px 1"
PASS element.style.columnWidth is "0px"
PASS element.style.columnCount is "1"
PASS element.style.columns is "0px 1"
PASS successfullyParsed is true

TEST COMPLETE
